requirements.txt
Flask==1.1.1
python-dateutil==2.8.0
pytz==2019.3
watchdog==0.10.1
sty==1.0.0b12
requests==2.22.0